+++ src/ast/passes/semantic_analyser.cpp
@@ -346,6 +346,21 @@ void SemanticAnalyser::visit(Builtin &builtin)
         LOG(ERROR, builtin.loc, err_)
             << "The func builtin can not be used with '"
             << attach_point->provider << "' probes";
+
+      if (type == ProbeType::uretprobe) {
+        // Disabled due to #2285
+        LOG(ERROR, builtin.loc, err_)
+            << "The 'func' builtin is not available for uretprobes. Consider "
+               "using 'probe' instead.";
+      }
+
+      if (type == ProbeType::kretprobe &&
+          !bpftrace_.feature_->has_helper_get_func_ip()) {
+        LOG(ERROR, builtin.loc, err_)
+            << "The 'func' builtin is not available for kretprobes on older "
+               "kernels without the get_func_ip BPF helper. Consider using "
+               "'probe' instead.";
+      }
     }
   } else if (!builtin.ident.compare(0, 3, "arg") && builtin.ident.size() == 4 &&
              builtin.ident.at(3) >= '0' && builtin.ident.at(3) <= '9') {